Extract document colors logic into a separate module
