+++ crates/editor/src/display_map/inlay_map.rs
@@ -305,7 +305,7 @@ impl<'a> Iterator for InlayChunks<'a> {
                     InlayId::Hint(_) => self.highlight_styles.inlay_hint,
                     InlayId::DebuggerValue(_) => self.highlight_styles.inlay_hint,
                     InlayId::Color(_) => {
-                        // // TODO kb: various modes and styles for inlays
+                        // TODO kb: various modes and styles for inlays
                         self.highlight_styles.inlay_hint
                     }
                 };
